래버리지 적용 api 수정

+++ src/main/java/com/corinne/corinne_be/dto/candle_dto/DateReponseDto.java
@@ -6,14 +6,14 @@ import lombok.Getter;
 public class DateReponseDto {
 
     private String tiker;
-    private int tradePrice;
+    private Long tradePrice;
     private double fluctuationRate;
 
 
     public DateReponseDto() {
     }
 
-    public DateReponseDto(String tiker, int tradePrice, double fluctuationRate) {
+    public DateReponseDto(String tiker, Long tradePrice, double fluctuationRate) {
         this.tiker = tiker;
         this.tradePrice = tradePrice;
         this.fluctuationRate = fluctuationRate;
